+++ packages/create-yoshi-app/__tests__/generateProject.spec.js
@@ -1,8 +1,8 @@
 const fs = require('fs');
 const path = require('path');
 const tempy = require('tempy');
-const globby = require('globby');
 const { generateProject } = require('../src');
+const getFilesInDir = require('../src/getFilesInDir');
 
 const projectTypes = fs.readdirSync(path.join(__dirname, '../templates'));
 
@@ -13,7 +13,7 @@ const mockedAnswers = projectTypes.map(projectType => [
     authorName: 'rany',
     authorEmail: 'rany@wix.com',
     organization: 'wix',
-    projectType: projectType,
+    projectType: projectType.replace('-typescript', ''),
     transpiler: projectType.endsWith('typescript') ? 'typescript' : 'babel',
   },
 ]);
@@ -23,18 +23,7 @@ test.each(mockedAnswers)('%s', (testName, answers) => {
 
   generateProject(answers, tempDir);
 
-  const filesPaths = globby.sync('**/*', {
-    cwd: tempDir,
-    dot: true,
-    gitignore: true,
-  });
-
-  const files = {};
-
-  filesPaths.forEach(filePath => {
-    const content = fs.readFileSync(path.join(tempDir, filePath), 'utf-8');
-    files[filePath] = content;
-  });
+  const files = getFilesInDir(tempDir);
 
   expect(files).toMatchSnapshot();
 });
